Løse funksjoner og structs flyttet ut til egen dbwriter pakke
